+++ src/gradescopeapi/__main__.py
@@ -1,9 +1,5 @@
-# Example showing how to import different components of the package and use them in the main function
-
-
 def main():
     pass
-
 
 if __name__ == "__main__":
     main()
