+++ src/gradescopeapi/classes/_courses.py
