+++ src/gradescopeapi/classes/account.py
@@ -2,14 +2,14 @@ from bs4 import BeautifulSoup
 from typing import List, Dict
 import time
 
-from gradescopeapi._classes._course_helpers import get_courses_info, get_course_members
-from gradescopeapi._classes._assignment_helpers import (
+from gradescopeapi.classes._helpers._course_helpers import get_courses_info, get_course_members
+from gradescopeapi.classes._helpers._assignment_helpers import (
     check_page_auth,
     get_assignments_instructor_view,
     get_assignments_student_view,
     get_submission_files,
 )
-from gradescopeapi._classes._assignments import Assignment
+from gradescopeapi.classes.assignments import Assignment
 
 
 class Account:
