+++ tests/test_connection.py
@@ -2,7 +2,7 @@ import requests
 import os
 from dotenv import load_dotenv
 
-from gradescopeapi._classes._login_helpers import (
+from gradescopeapi.classes._helpers._login_helpers import (
     login_set_session_cookies,
     get_auth_token_init_gradescope_session,
 )
